+++ src/test/resources/edgeCases/firstWithColumns.php
@@ -1,0 +1,7 @@
+<?php
+
+class User extends \Hyperf\Database\Model\Model {
+
+}
+
+User::query()->first(['id', '<caret>']);